Create project so we can use local library table
Create a specific project with empty schematic and layout (except for
some explanatory text). This is done in order to add a project specific
library tables for footprints and symbols. This will allow us to open
such libraries on any computer where the repository has been cloned
without having to setup environment variables prior to start editing
the component libraries.
3D model paths in footprint have been updated accordingly